Move authentication system to asovi_app
