--- accounts/admin.py
@@ -1,6 +1,0 @@
-from django.contrib import admin
-
-from .models import CustomUser
-
-
-admin.site.register(CustomUser)
