--- accounts/templates/account/email/email_confirmation_signup_message.txt
@@ -1,1 +1,0 @@
-{% include "account/email/email_confirmation_message.txt" %}
