--- accounts/templates/account/email/email_confirmation_signup_subject.txt
@@ -1,1 +1,0 @@
-{% include "account/email/email_confirmation_subject.txt" %}
