--- accounts/templates/account/email/email_confirmation_subject.txt
@@ -1,4 +1,0 @@
-{% load i18n %}
-{% autoescape off %}
-{% blocktrans %}Please Confirm Your E-mail Address{% endblocktrans %}
-{% endautoescape %}
