--- accounts/templates/account/email/password_reset_key_subject.txt
@@ -1,4 +1,0 @@
-{% load i18n %}
-{% autoescape off %}
-{% blocktrans %}Password Reset E-mail{% endblocktrans %}
-{% endautoescape %}
