+++ asovi_app/templates/asovi_app/email_confirm.html
